refacto segment
